requirements.txt
# Core Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Web Frameworks
streamlit>=1.32.0
flask==3.0.0
flask-cors==4.0.0

# Audio Processing & Speech Recognition
librosa>=0.10.0
SpeechRecognition>=3.10.0
openai-whisper>=20231117
soundfile>=0.12.1

# Image & Video Processing
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.33.0
scipy>=1.11.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Cloud APIs
google-cloud-speech>=2.20.0
huggingface-hub>=0.17.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
